added bug in the voxel morphology funcs added action #47

+++ include/delfem2/gridvoxel.cpp
@@ -283,19 +283,19 @@ int dfm2::Adj_Grid
 void delfem2::Grid3Voxel_Dilation
  (dfm2::CGrid3<int>& grid)
 {
-  const unsigned int nx = grid.ndivx;
-  const unsigned int ny = grid.ndivy;
-  const unsigned int nz = grid.ndivz;
-  for(unsigned int iz1=0;iz1<nz;++iz1){
-    for(unsigned int iy1=0;iy1<ny;++iy1){
-      for(unsigned int ix1=0;ix1<nx;++ix1){
+  const int nx = (int)grid.ndivx;
+  const int ny = (int)grid.ndivy;
+  const int nz = (int)grid.ndivz;
+  for(int iz1=0;iz1<nz;++iz1){
+    for(int iy1=0;iy1<ny;++iy1){
+      for(int ix1=0;ix1<nx;++ix1){
         int flgx0 = ix1-1>=0 ? grid.aVal[iz1*ny*nx + iy1*nx + (ix1-1)] : 0;
         int flgx2 = ix1+1<nx ? grid.aVal[iz1*ny*nx + iy1*nx + (ix1+1)] : 0;
         int flgy0 = iy1-1>=0 ? grid.aVal[iz1*ny*nx + (iy1-1)*nx + ix1] : 0;
         int flgy2 = iy1+1<ny ? grid.aVal[iz1*ny*nx + (iy1+1)*nx + ix1] : 0;
         int flgz0 = iz1-1>=0 ? grid.aVal[(iz1-1)*ny*nx + iy1*nx + ix1] : 0;
         int flgz2 = iz1+1<nz ? grid.aVal[(iz1+1)*ny*nx + iy1*nx + ix1] : 0;
-        const unsigned int ivox = iz1*ny*nx + iy1*nx + ix1;
+        const int ivox = iz1*ny*nx + iy1*nx + ix1;
         if( grid.aVal[ivox] != 0 ) continue;
         if( flgx0 == 1 ){ grid.aVal[ivox] = 2; }
         if( flgx2 == 1 ){ grid.aVal[ivox] = 2; }
@@ -306,8 +306,8 @@ void delfem2::Grid3Voxel_Dilation
       }
     }
   }
-  const unsigned int nvox = nx*ny*nz;
-  for(unsigned int ivox=0;ivox<nvox;++ivox){
+  const int nvox = nx*ny*nz;
+  for(int ivox=0;ivox<nvox;++ivox){
     if( grid.aVal[ivox] == 0 ){ continue; }
     grid.aVal[ivox] = 1;
   }
@@ -316,19 +316,19 @@ void delfem2::Grid3Voxel_Dilation
 void delfem2::Grid3Voxel_Erosion
 (dfm2::CGrid3<int>& grid)
 {
-  const unsigned int nx = grid.ndivx;
-  const unsigned int ny = grid.ndivy;
-  const unsigned int nz = grid.ndivz;
-  for(unsigned int iz1=0;iz1<nz;++iz1){
-    for(unsigned int iy1=0;iy1<ny;++iy1){
-      for(unsigned int ix1=0;ix1<nx;++ix1){
+  const int nx = (int)grid.ndivx;
+  const int ny = (int)grid.ndivy;
+  const int nz = (int)grid.ndivz;
+  for(int iz1=0;iz1<nz;++iz1){
+    for(int iy1=0;iy1<ny;++iy1){
+      for(int ix1=0;ix1<nx;++ix1){
         int flgx0 = ix1-1>=0 ? grid.aVal[iz1*ny*nx + iy1*nx + (ix1-1)] : 0;
         int flgx2 = ix1+1<nx ? grid.aVal[iz1*ny*nx + iy1*nx + (ix1+1)] : 0;
         int flgy0 = iy1-1>=0 ? grid.aVal[iz1*ny*nx + (iy1-1)*nx + ix1] : 0;
         int flgy2 = iy1+1<ny ? grid.aVal[iz1*ny*nx + (iy1+1)*nx + ix1] : 0;
         int flgz0 = iz1-1>=0 ? grid.aVal[(iz1-1)*ny*nx + iy1*nx + ix1] : 0;
         int flgz2 = iz1+1<nz ? grid.aVal[(iz1+1)*ny*nx + iy1*nx + ix1] : 0;
-        const unsigned int ivox = iz1*ny*nx + iy1*nx + ix1;
+        const int ivox = iz1*ny*nx + iy1*nx + ix1;
         if( grid.aVal[ivox] == 0 ) continue;
         if( flgx0 == 0 ){ grid.aVal[ivox] = 2; }
         if( flgx2 == 0 ){ grid.aVal[ivox] = 2; }
@@ -339,8 +339,8 @@ void delfem2::Grid3Voxel_Erosion
       }
     }
   }
-  const unsigned int nvox = nx*ny*nz;
-  for(unsigned int ivox=0;ivox<nvox;++ivox){
+  const int nvox = nx*ny*nz;
+  for(int ivox=0;ivox<nvox;++ivox){
     if( grid.aVal[ivox] != 2 ){ continue; }
     grid.aVal[ivox] = 0;
   }
